src/nlp-based-preprocessing/requirements.txt
pandas>=2.2
pyyaml>=6.0
pydantic>=2.8
SQLAlchemy>=2.0
psycopg2-binary>=2.9
tabulate>=0.9
openpyxl>=3.1
geopandas>=1.0 ; platform_system!="Windows" or python_version>="3.11"
